WIP:
Updated the list of years to point to an array of constant values instead of an observable array. This keeps the list from updating whenever the list is filtered.
Updated the filterByMonth and filterByYear methods to filter on the month or year the pdf was released instead of the label.
Updated the month service to provide the numerical value for each month in order to match what's returned by the server for each record.